requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Caching==2.1.0
gunicorn==21.2.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
Werkzeug==2.3.7
python-dateutil==2.8.2
pytz==2023.3.post1
joblib==1.3.2 